+++ src/main/java/com/getpcpanel/util/FileUtil.java
@@ -4,23 +4,26 @@ import java.io.File;
 
 import javax.annotation.PostConstruct;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
 @Service
+@RequiredArgsConstructor
 public class FileUtil {
-    public static final File FILES_ROOT = new File(System.getProperty("user.home"), ".pcpanel");
+    @Value("${application.root}") private final File root;
 
     @PostConstruct
     void ensureRoot() {
-        if (!FILES_ROOT.exists() && !FILES_ROOT.mkdirs()) {
-            log.error("Unable to create file root: {}", FILES_ROOT);
+        if (!root.exists() && !root.mkdirs()) {
+            log.error("Unable to create file root: {}", root);
         }
     }
 
     public File getFile(String file) {
-        return new File(FILES_ROOT, file);
+        return new File(root, file);
     }
 }