+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 **New:**
 
+- When there is no saved profile but there is one from the original software, ask to migrate on startup
 - Right clicking a knob or slider opens the lighting dialog
     - Middle click already triggered the click action (I did not know that)
+
+**Fix:**
+
+- Its possible to configure an alternative profile folder for development purposes
 
 ## [1.2]